requirements.txt
metatrader5
websockets
pandas
aiohttp
numpy
python-dotenv
